# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Market Pulse API",
    "description": "Job market intelligence API providing tech job listings and salary benchmarks. Search jobs from The Muse and Adzuna by title, company, location, category, or level. Compare compensation across cities, experience levels, and companies.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://market-pulse-app.25raqj64rgw0.us-south.codeengine.appdomain.cloud"
    }
  ],
  "paths": {
...
    "/salaries/search": {
      "get": {
        "tags": [
          "Salaries"
        ],
        "summary": "Search salary data from Adzuna job postings",
        "description": "Aggregates salary information from Adzuna job postings that include salary data. Returns individual salary records plus aggregate stats (min, max, median, average). Requires a job title; location and company are optional filters. All text filters are case-insensitive partial matches.",
        "operationId": "search_salaries",
        "parameters": [
          {
            "name": "job_title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Job title keyword, e.g. 'frontend developer', 'data scientist'",
              "title": "Job Title"
            },
            "description": "Job title keyword, e.g. 'frontend developer', 'data scientist'"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "City, state, or region, e.g. 'California', 'New York'",
              "title": "Location"
            },
            "description": "City, state, or region, e.g. 'California', 'New York'"
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Company name, e.g. 'TikTok', 'Amazon'",
              "title": "Company"
            },
            "description": "Company name, e.g. 'TikTok', 'Amazon'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max individual salary records to return",
              "default": 25,
              "title": "Limit"
            },
            "description": "Max individual salary records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SalaryAggregation": {
        "properties": {
          "avg_salary_min": {
            "type": "number",
            "title": "Avg Salary Min",
            "description": "Average of salary_min across matches"
          },
          "avg_salary_max": {
            "type": "number",
            "title": "Avg Salary Max",
            "description": "Average of salary_max across matches"
          },
          "overall_min": {
            "type": "number",
            "title": "Overall Min",
            "description": "Lowest salary_min"
          },
          "overall_max": {
            "type": "number",
            "title": "Overall Max",
            "description": "Highest salary_max"
          },
          "median_salary": {
            "type": "number",
            "title": "Median Salary",
            "description": "Median of (salary_min + salary_max) / 2"
          }
        },
        "type": "object",
        "required": [
          "avg_salary_min",
          "avg_salary_max",
          "overall_min",
          "overall_max",
          "median_salary"
        ],
        "title": "SalaryAggregation",
        "description": "Aggregate salary statistics across matching job postings."
      },
      "SalaryRecord": {
        "properties": {
          "job_title": {
            "type": "string",
            "title": "Job Title",
            "description": "Job title"
          },
          "company": {
            "type": "string",
            "title": "Company",
            "description": "Company name"
          },
          "location": {
            "type": "string",
            "title": "Location",
            "description": "Job location"
          },
          "salary_min": {
            "type": "number",
            "title": "Salary Min",
            "description": "Minimum salary in USD/year"
          },
          "salary_max": {
            "type": "number",
            "title": "Salary Max",
            "description": "Maximum salary in USD/year"
          }
        },
        "type": "object",
        "required": [
          "job_title",
          "company",
          "location",
          "salary_min",
          "salary_max"
        ],
        "title": "SalaryRecord",
        "description": "A single salary data point from an Adzuna job posting."
      },
      "SalaryResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of matching salary records"
          },
          "aggregation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SalaryAggregation"
              },
              {
                "type": "null"
              }
            ],
            "description": "Aggregate stats (null if no results)"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/SalaryRecord"
            },
            "type": "array",
            "title": "Records"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Hint when no results are found"
          }
        },
        "type": "object",
        "required": [
          "count",
          "records"
        ],
        "title": "SalaryResponse",
        "description": "Salary search results with aggregation."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}